Fix failing tests
PollingJobActivator tests needed to be updated with regard to earlier
changes where Dispose is now called on a PollingJob instead of Stop.

InMemoryQueueIntegrationTester needed to use a different way to see if
the Received queue receives the message. You can't use the .Any() LINQ
method because it's an extension method on IEnumerable and you are can't
enumerate collections that change. Using collection count (length)
seemed to be the right way to get around that problem.